Move agent process tracking and cleanup into agent-specific util package
Move agent S3 utilities into agent-specific util package
