+++ agent/logging.go
@@ -10,10 +10,10 @@ import (
 	"strings"
 
 	"github.com/evergreen-ci/evergreen"
+	"github.com/evergreen-ci/evergreen/agent/util"
 	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/rest/client"
-	"github.com/evergreen-ci/evergreen/util"
 	"github.com/evergreen-ci/pail"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/level"
@@ -238,12 +238,12 @@ func (a *Agent) uploadLogDir(ctx context.Context, tc *taskContext, bucket pail.B
 	return catcher.Resolve()
 }
 
-func (a *Agent) uploadSingleFile(ctx context.Context, tc *taskContext, bucket pail.Bucket, file string, taskID string, execution int, command string) error {
+func (a *Agent) uploadSingleFile(ctx context.Context, tc *taskContext, bucket pail.Bucket, file string, taskID string, execution int, cmd string) error {
 	localDir := filepath.Join(a.opts.WorkingDirectory, taskLogDirectory)
 	remotePath := fmt.Sprintf("logs/%s/%s", taskID, strconv.Itoa(execution))
-	if command != "" {
-		localDir = filepath.Join(localDir, command)
-		remotePath = fmt.Sprintf("%s/%s", remotePath, command)
+	if cmd != "" {
+		localDir = filepath.Join(localDir, cmd)
+		remotePath = fmt.Sprintf("%s/%s", remotePath, cmd)
 	}
 	localPath := filepath.Join(localDir, file)
 	_, err := os.Stat(localPath)
@@ -259,17 +259,17 @@ func (a *Agent) uploadSingleFile(ctx context.Context, tc *taskContext, bucket pa
 	switch file {
 	case agentLogFileName:
 		tc.logs.AgentLogURLs = append(tc.logs.AgentLogURLs, apimodels.LogInfo{
-			Command: command,
+			Command: cmd,
 			URL:     remoteURL,
 		})
 	case systemLogFileName:
 		tc.logs.SystemLogURLs = append(tc.logs.SystemLogURLs, apimodels.LogInfo{
-			Command: command,
+			Command: cmd,
 			URL:     remoteURL,
 		})
 	case taskLogFileName:
 		tc.logs.TaskLogURLs = append(tc.logs.TaskLogURLs, apimodels.LogInfo{
-			Command: command,
+			Command: cmd,
 			URL:     remoteURL,
 		})
 	}
