+++ agent/util/subtree.go
