+++ agent/util/subtree_windows.go
