+++ makefile
@@ -3,7 +3,7 @@ name := evergreen
 buildDir := bin
 tmpDir := $(abspath $(buildDir)/tmp)
 nodeDir := public
-packages := $(name) agent operations cloud cloud-userdata command db util plugin units graphql
+packages := $(name) agent agent-util operations cloud cloud-userdata command command-util db util plugin units graphql
 packages += thirdparty auth scheduler model validator service repotracker
 packages += model-patch model-artifact model-host model-build model-event model-task model-user model-distro model-manifest model-testresult
 packages += operations-metabuild-generator operations-metabuild-model
