+++ self-tests.yml
@@ -417,6 +417,9 @@ tasks:
   - <<: *run-go-test-suite-with-mongodb
     tags: ["db", "test", "agent"]
     name: test-agent
+  - <<: *run-go-test-suite
+    tags: ["nodb", "test", "agent"]
+    name: test-agent-util
   - <<: *run-go-test-suite-with-mongodb
     tags: ["db", "test"]
     name: test-rest-data
@@ -454,7 +457,7 @@ tasks:
     tags: ["db", "test"]
     name: test-trigger
   - <<: *run-go-test-suite
-    tags: ["nodb", "test", "agent"]
+    tags: ["nodb", "test"]
     name: test-util
   - <<: *run-go-test-suite-with-mongodb
     tags: ["db", "test"]
